поправил мапинг, доработал бизнес логику

+++ orderService/src/main/java/ru/aston/dto/OrdersDto.java
@@ -2,6 +2,7 @@ package ru.aston.dto;
 
 import ru.aston.entity.OrderStatus;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -10,7 +11,7 @@ public record OrdersDto(
         Long userId,
         Long restaurantId,
         List<OrderItemDto> items,
-        Double totalPrice,
+        BigDecimal totalPrice,
         OrderStatus status,
         LocalDateTime createdAt
 ) {}
